telink: kconfig: soc: provide a single config for ot rcp buffers
W91 OT CLI works unstable with a 64 byte RX ISR buffer.
This may be related to some issue with RTS and CTS signals
and should be additionally verified. As a temporary solution,
the ISR and general OT RCP buffer are set to the same size (256 bytes)
OT_RCP_RX_ISR_BUFFER_SIZE is removed because it is not currently used.

Signed-off-by: Nazarii Bodnar <nazarii.bodnar@telink-semi.com>